fix(cli-sub-agent): align consolidated findings filename across writer + reader (#791)
The multi-reviewer consolidated artifact writer emitted a filename that no longer matched the prior-round reader path handling. Reader never matched, so cumulative review's "Prior-Round Assumptions to Re-verify" silently dropped structured findings in multi-reviewer mode.

Reader cleanup now relies on the single consolidated artifact path only and drops the dead review-findings.json fallback. A regression test covers the consolidated writer-to-reader prior-round path end to end.

Rule 019 (pre-production) - no back-compat alias for the old name.

Closes #791.
